+++ pydl/pydlspec2d/spec1d/pca_gal.py
@@ -10,17 +10,17 @@ def pca_gal(**kwargs):
 
     Parameters
     ----------
-    inputfile : str, optional
+    inputfile : :class:`str`, optional
         The list of spectra to use.  If not specified, $IDLSPEC2D_DIR/tempates/eigeninput_gal.dat will be used.
-    wavemin : float, optional
+    wavemin : :class:`float`, optional
         Minimum wavelength for the template.  If not specified 1850 Å will be used.
-    wavemax : float, optional
+    wavemax : :class:`float`, optional
         Maximum wavelength for the template.  If not specified 10000 Å will be used.
-    niter : int, optional
+    niter : :class:`int`, optional
         Number of iterations.  The default is 10.
-    dump : str, optional
+    dump : :class:`str`, optional
         If set, save input data in a Python pickle file.
-    flux : bool, optional
+    flux : :class:`bool`, optional
         If set to ``True`` make some additional QA plots of the input spectra.
         The default is ``False``.