+++ pydl/pydlutils/yanny/write_ndarray_to_yanny.py
@@ -1,8 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
-from astropy.extern import six
-
-
 def write_ndarray_to_yanny(filename,datatables,structnames=None,
                            enums=None,hdr=None,comments=None):
     """Converts a NumPy record array into a new FTCL/yanny file.
@@ -11,29 +8,32 @@ def write_ndarray_to_yanny(filename,datatables,structnames=None,
 
     Parameters
     ----------
-    filename : str
+    filename : :class:`str`
         The name of a parameter file.
-    datatables : numpy.ndarray or list of numpy.ndarray
-        A NumPy record array containing data that can be copied into a yanny object.
-    structnames : str or list of str, optional
+    datatables : :class:`numpy.ndarray`, :class:`numpy.recarray` or :class:`list` of these.
+        A NumPy record array containing data that can be copied into a `yanny` object.
+    structnames : :class:`str` or :class:`list` of :class:`str`, optional
         The name(s) to give the structure(s) in the yanny file.  Defaults to 'MYSTRUCT0'.
-    enums : dict, optional
+    enums : :class:`dict`, optional
         A dictionary containing enum information.  See the documentation for
-        the `dtype_to_struct` method of the yanny object.
-    hdr : dict, optional
+        the :meth:`~pydl.pydlutils.yanny.yanny.dtype_to_struct` method of the yanny object.
+    hdr : :class:`dict`, optional
         A dictionary containing keyword/value pairs for the 'header' of the yanny file.
-    comments : str or list of str, optional
+    comments : :class:`str` or :class:`list` of :class:`str`, optional
         A string containing comments that will be added to the start of the new file.
 
     Returns
     -------
-    par : pydl.pydlutils.yanny.yanny
+    par : `yanny`
         The `yanny` object resulting from writing the file.
 
-    Examples
-    --------
+    Raises
+    ------
+    PydlutilsException
+        If `filename` already exists, or if the metadata are incorrect.
     """
-    from numpy import ndarray
+    from numpy import ndarray, recarray
+    from astropy.extern.six import string_types
     from . import yanny
     from .. import PydlutilsException
     par = yanny(filename,np=True)
@@ -42,11 +42,11 @@ def write_ndarray_to_yanny(filename,datatables,structnames=None,
         # If the file already exists
         #
         raise PydlutilsException("Apparently {0} already exists.".format(filename))
-    if type(datatables) == ndarray:
+    if isinstance(datatables, (ndarray,recarray)):
         datatables = (datatables,)
     if structnames is None:
         structnames = ["MYSTRUCT{0:d}".format(k) for k in range(len(datatables))]
-    if isinstance(structnames, six.string_types):
+    if isinstance(structnames, string_types):
         structnames = (structnames,)
     if len(datatables) != len(structnames):
         raise PydlutilsException("The data tables and their names do not match!")